# api/docs/swagger.yaml
openapi: 3.0.0
info:
  title: Edwardsville Memorial Tracker API
  description: REST API for Edwardsville Memorial Tracker
  version: 1.0.0
  license:
    name: MIT

servers:
  - url: http:localhost:1337/api

paths:
  /memorials:
    tags:
    - Memorials
    get:
      summary: Get all memorials
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Memorial"
    post:
      summary: Add new memorial
      responses:
        "201":
          description: Created
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Memorial"
  /memorials/{id}:
    delete:
      summary: Delete memorial by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /types:
    get:
      summary: Get all types
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Type"
    post:
      summary: Add new type
      responses:
        "201":
          description: Created
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
  /types/{id}:
    delete:
      summary: Delete type by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted

components:
  schemas:
    Memorial:
      type: object
      properties:
        id:
          type: string
        typeId:
          type: string

    Type:
      type: object
      properties:
        id:
          type: string
        name:
          type: string